requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.1.31
cffi
charset-normalizer==2.0.12
click==8.0.4
cryptography
cx_Oracle==8.3.0
dataclasses==0.6
Flask==2.0.3
idna==3.7
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==5.3.1
MarkupSafe==2.0.1
oracledb
pip-chill==1.0.3
prometheus-client==0.17.1
pycparser
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.19
Werkzeug==2.0.3
zipp==3.6.0
